Fix build break introduced by r53882.
CR:  dalmeida


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@53885 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ lsass/tools/lw-lsa/ad_get_machine.c
@@ -160,7 +160,7 @@ PrintAccountInfo(
            LW_PRINTF_STRING(pAccountInfo->Fqdn),
            pAccountInfo->Type,
            pAccountInfo->KeyVersionNumber,
-           pAccountInfo->LastChangeTime);
+           (long long int)pAccountInfo->LastChangeTime);
 }
 
 static
